Install and configure TailwindCSS
