+++ src/constants/index.js
@@ -1,0 +1,9 @@
+const links = [
+  { path: '/', text: 'Houses' },
+  { path: '/reserve', text: 'Reserve' },
+  { path: '/myReservations', text: 'My Reservations' },
+  { path: '/addHouse', text: 'Add House' },
+  { path: '/deleteHouse', text: 'Delete House' },
+];
+
+export default links;
